feat: detect duplicate-named universes on the Universes page with inline merge (#493)
Surfaces the existing duplicate-resolution system directly on the Universes
index so a duplicate that slipped past sync time can be merged without
visiting Sharing → Duplicates. When two universes share a name, a warning
banner lists each group with Merge / Rename / Keep-both controls.

To avoid duplicating the merge UI, the flow is extracted from DuplicatesTab
into reusable pieces shared by both surfaces:
- useRecordMerge hook (open → dry-run preview → resolve conflicts → execute,
  with an onMerged refresh callback)
- MergeModal (survivor picker + per-field conflict resolver + cascade summary)
- DuplicateGroup + RecordRow (per-group rows with inline rename)

DuplicatesTab is rewired onto these with identical behavior. The Universes
page fetches listUniverseDuplicates on mount and re-scans after a merge or
rename (the loser is tombstoned, child series re-pointed, media folded).
