+++ .changelog/NEXT.md
@@ -10,7 +10,7 @@ TBD
 
 - **Writers Room Guide.** A new deep-linkable documentation page at `/writers-room/guide` (reachable via a "Guide" link in the Writers Room header and from `⌘K` / voice). It documents the literary length ladder — Microfiction, Flash Fiction, Standard Short Story, Novelette, plus Novella/Novel for context — with both word and character bands, a page-based book-length estimate table (200/300 pages), and craft principles grouped by structure, character, prose, and revision. The length targets and principles live in a single canonical data module (`client/src/lib/writingGuide.js`, with a `classifyByWordCount()` helper) so forthcoming editor analyses — including the planned **emotional-roadmap evaluator** that charts the reader's emotional journey beat by beat — read the same source the docs render from.
 - **Sync hygiene: duplicate resolution + a non-blocking edit-conflict journal.** Because two PortOS installs (or two machines federated as peers, or two copies of a `data/` folder synced through Google Drive/Dropbox) can each independently create a Universe or Series with the same name — and records are keyed by random id, not name — sync used to leave silent duplicates (e.g. two "Clandestiny" universes) and silently discard the loser whenever both sides edited the same record. Two new tabs under **Sharing** fix this:
-  - **Duplicates** — finds same-named-but-different-id Universes (and Series, scoped within their universe) and offers to **Merge** (smart field-union: canon, categories, variations, influences, and render history are unioned; genuinely-conflicting fields are resolved field-by-field via a side-by-side diff; child series/issues and media collections are re-pointed to the survivor; the loser is tombstoned and propagated to peers) or **Rename** one to keep both.
+  - **Duplicates** — finds same-named-but-different-id Universes (and Series, scoped within their universe) and offers to **Merge** (smart field-union: canon, categories, variations, influences, and render history are unioned; genuinely-conflicting fields are resolved field-by-field via a side-by-side diff; child series/issues and media collections are re-pointed to the survivor; the loser is tombstoned and propagated to peers) or **Rename** one to keep both. The same detection + merge is also surfaced **inline on the Universes page** — when two universes share a name a warning banner appears at the top of the list with the same Merge / Rename / Keep-both controls, so a duplicate that slipped past sync time can be resolved without leaving the page.
   - **Conflicts** — when a cross-install last-write-wins merge would overwrite a record both sides edited, the losing version is now archived to a local conflict journal instead of vanishing. Review the diff and **Restore mine**, **Merge** specific fields back, or **Discard**. Convergence is never blocked — peers still agree on the LWW winner; no edit is lost.
 
   Both surfaces are deep-linkable (`/sharing/duplicates`, `/sharing/conflicts`) and reachable from ⌘K / voice.
