+++ client/src/hooks/index.js
@@ -76,6 +76,7 @@ export * from './useAppOperation.js';
 export * from './useCityData.js';
 export * from './useDeathClock.js';
 export * from './usePostSession.js';
+export * from './useRecordMerge.js';
 export * from './useSyncIntegrity.js';
 export * from './useUniverseNav.js';
 export * from './useVoiceUiSync.js';
